Move all Integration Test 'Provider::Admin::Account::UsersControllerTest' to the same file inside the right folder
